Moved functionality to a module
